feat(glusterfs): nodes mount storage on boot
This commit is a bit ugly, there are some unrelated changed pulled in
due to the early nature of this project.

Current I am not running separate consul master nodes, so the following
assumptions are made.

- All three nodes are initially up, this allows the consul leader
  election to happen.
- Once this happens, glusterfs is configured on two nodes.
- Once two nodes are good, a third node is configured
- if any single node goes down, drop the cluster to 2 nodes
- if a third node comes up again, bring it into the cluster

I need to handle the case where two nodes go down, but that could become
more complicated, and I may have to delete and recreate the replica set
if I am unable to have a replica set size of 1.